pyproject.toml
[tool.poetry]
name = "coursework-7"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "coursework_7"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
stripe = "^8.0.0"
redis = "^5.0.1"
celery = "^5.3.6"
eventlet = "^0.35.1"
telebot = "^0.0.5"
flake8 = "^7.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
